% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_quiz.R
\name{activate_quiz}
\alias{activate_quiz}
\title{Activate quiz}
\usage{
activate_quiz(
  course_id,
  quiz_id,
  unlock_at,
  time_open = minutes(10),
  access_code
)
}
\arguments{
\item{course_id}{id number of the course.
Go to your course home page and the number is at the end of the url.}

\item{quiz_id}{The id number of the quiz.
This is the number at the end of the quiz url.}

\item{unlock_at}{date-time to unlock quiz - defaults to current time
(Canvas accepts times in ISO 8601 format, e.g. 2011-10-21T18:48Z).
You can use \code{\link[=iso8601]{iso8601()}} to format the date-time correctly.}

\item{time_open}{time to keep quiz open (in minutes) -
default to ten minutes.
Set to NULL to keep quiz open}

\item{access_code}{Optional access code for the quiz
(use as an alternative to setting a passcode within the quiz).}
}
\value{
A quiz object (invisible)
}
\description{
Activates pre-existing quiz (perhaps made with \code{\link[=create_quiz]{create_quiz()}})
by setting unlock time, due time and publishes it.
}
