% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_passcode.R
\docType{data}
\name{adjective_noun}
\alias{adjective_noun}
\title{A list with adjective and nouns}
\format{
A list
\describe{
\item{norsk}{A list of Norwegian adjectives and nouns
(mammals and birds) without special characters.}
\item{english}{A list of English adjectives and nouns.}
}
}
\usage{
data(adjective_noun)
}
\description{
A list with adjective and nouns
}
\keyword{datasets}
