% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_params.R
\name{collate_question_params}
\alias{collate_question_params}
\title{Collate quiz questions parameters}
\usage{
collate_question_params(
  question_name = "qn",
  question_text = "Please enter the passcode",
  question_type = "short_answer_question",
  position = 1,
  correct_comments = "Thankyou",
  incorrect_comments = "Please check your passcode carefully",
  answers = NULL
)
}
\arguments{
\item{question_name}{Name of the question}

\item{question_text}{Question text}

\item{question_type}{Type of question}

\item{position}{Integer setting question order}

\item{correct_comments}{Response for a correct answer}

\item{incorrect_comments}{Response for an incorrect answer}

\item{answers}{The correct answer}
}
\value{
A list
}
\description{
Collates quiz question parameters.
}
\details{
More information about each parameter from the \href{https://developerdocs.instructure.com/services/canvas/resources/quiz_questions#method.quizzes/quiz_questions.create}{canvas API documentation}
}
