% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_register.R
\name{create_attendence}
\alias{create_attendence}
\title{Create attendance quiz and qr code}
\usage{
create_attendence(
  course_id,
  passcode = random_passcode(adjective_noun_list = adjective_noun$norsk),
  quiz_params = collate_quiz_params()
)
}
\arguments{
\item{course_id}{id number of the course.
Go to your course home page and the number is at the end of the url.}

\item{passcode}{the passcode}

\item{quiz_params}{Quiz parameters.
You can make these with \code{\link[=collate_quiz_params]{collate_quiz_params()}}}
}
\description{
Create attendance quiz and qr code
}
