% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_canvas.R
\name{create_quiz}
\alias{create_quiz}
\title{Create quiz}
\usage{
create_quiz(canvas, course_id, quiz_params)
}
\arguments{
\item{canvas}{Authentication using \code{\link[vvcanvas:canvas_authenticate]{vvcanvas::canvas_authenticate()}}}

\item{course_id}{id number of the course.
Go to your course home page and the number is at the end of the url.}

\item{quiz_params}{Quiz parameters.
You can make these with \code{\link[=collate_quiz_params]{collate_quiz_params()}}}
}
\value{
A list with a canvas quiz.
}
\description{
Create a canvas quiz
}
