% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_canvas.R
\name{create_quiz_question}
\alias{create_quiz_question}
\title{Create quiz question}
\usage{
create_quiz_question(canvas, course_id, quiz_id, quiz_question_params)
}
\arguments{
\item{canvas}{Authentication using \code{\link[vvcanvas:canvas_authenticate]{vvcanvas::canvas_authenticate()}}}

\item{course_id}{id number of the course.
Go to your course home page and the number is at the end of the url.}

\item{quiz_id}{The id number of the quiz.
You can extract this from the output of \code{\link[=create_quiz]{create_quiz()}}.}

\item{quiz_question_params}{Quiz question parameters.
You can make these with \code{\link[=collate_question_params]{collate_question_params()}}}
}
\value{
A list with a canvas quiz.
}
\description{
Create a canvas quiz question
}
