% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_register.R
\name{create_register}
\alias{create_register}
\title{Create attendance quiz and qr code}
\usage{
create_register(
  course_id,
  accesscode = random_passcode(adjective_noun_list = adjective_noun$norsk),
  quiz_params = digit_code(),
  quiz_id
)
}
\arguments{
\item{course_id}{id number of the course.
Go to your course home page and the number is at the end of the url.}

\item{accesscode}{the accesscode. Defaults to a four digit code.}

\item{quiz_params}{Quiz parameters.
You can make these with \code{\link[=collate_quiz_params]{collate_quiz_params()}}}

\item{quiz_id}{Optional id of an existing quiz to activate.}
}
\description{
Create attendance quiz and qr code
}
