% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_params.R
\name{iso8601}
\alias{iso8601}
\title{ISO8601}
\usage{
iso8601(datetime, offset)
}
\arguments{
\item{datetime}{Date-time object (POSIXct class) for
example made with \code{\link[lubridate:now]{lubridate::now()}}.}

\item{offset}{Offset period for example made with \code{\link[lubridate:minutes]{lubridate::minutes()}}.}
}
\description{
Formats date in ISO8601 date time format with timezone.
}
