% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_attendance.R
\name{plot_attendance}
\alias{plot_attendance}
\title{Plot course attendance}
\usage{
plot_attendance(attendance, lowest = 20)
}
\arguments{
\item{attendance}{Output of \code{get_course_attendance()}.}

\item{lowest}{Show the n students with the lowest attendance (to prevent the plot being overwhelming). Use \code{Inf} to show all students.}
}
\description{
Plot course attendance
}
