% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qrcode.R
\name{plot_qrcode}
\alias{plot_qrcode}
\title{Plot qrcode with passcode}
\usage{
plot_qrcode(x, col = c("white", "black"), main = "title", passcode)
}
\arguments{
\item{x}{qrcode}

\item{col}{colours for the qrcode}

\item{main}{Plot title}

\item{passcode}{passcode}
}
\description{
Plot qrcode with passcode
}
