% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_passcode.R
\name{random_passcode}
\alias{random_passcode}
\title{Make a random passcode}
\usage{
random_passcode(adjective_noun_list = adjective_noun$norsk)
}
\arguments{
\item{adjective_noun_list}{A list with adjectives and nouns.
See \link{adjective_noun}.}
}
\value{
A length-one character vector.
}
\description{
Makes a random passcode by sampling a list of adjectives and nouns
}
\details{
Random adjective-noun combinations are used to generate user names or repo
names by github and reddit.
}
