% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{today}
\alias{minutes}
\alias{glue}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{glue}{\code{\link[glue:glue]{glue()}}}

  \item{lubridate}{\code{\link[lubridate:minutes]{minutes()}}, \code{\link[lubridate:today]{today()}}}
}}

