% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_attendance.R
\name{summarise_attendance}
\alias{summarise_attendance}
\title{Summarise course attendance}
\usage{
summarise_attendance(attendance)
}
\arguments{
\item{attendance}{Output of \code{get_course_attendance()}.}
}
\description{
Summarise course attendance
}
