% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_passcode.R
\name{adjective_noun}
\alias{adjective_noun}
\title{A list with adjective and nouns}
\format{
A list
\describe{
\item{norsk}{A list of Norwegian adjectives and nouns
(mammals and birds) without special characters.}
\item{english}{A list of English adjectives and nouns.}
}
}
\description{
A list with adjective and nouns
}
