% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_params.R
\name{collate_quiz_params}
\alias{collate_quiz_params}
\title{Collate quiz parameters}
\usage{
collate_quiz_params(
  title = glue("Registration {today()}"),
  description = "Please press submit to register attendence",
  quiz_type = "graded_survey",
  unlock_at = NULL,
  due_at = NULL,
  ...
)
}
\arguments{
\item{title}{Title of the quiz. Defaults to "Registration \{Date\}"}

\item{description}{Description text}

\item{quiz_type}{Type of quiz}

\item{unlock_at}{Optional. Set using \code{\link[=iso8601]{iso8601()}}.}

\item{due_at}{Optional. Set using \code{\link[=iso8601]{iso8601()}}.}

\item{\dots}{Extra parameters. See details}
}
\value{
A list
}
\description{
Collates and processes quiz parameters.
}
\details{
More information about each parameter from the
\href{https://developerdocs.instructure.com/services/canvas/resources/quizzes#method.quizzes/quizzes_api.create}{canvas API documentation}
}
