% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_attendance.R
\name{get_course_attendance}
\alias{get_course_attendance}
\title{Get course attendance}
\usage{
get_course_attendance(course_id, pattern = "^Registration")
}
\arguments{
\item{course_id}{id number of the course.
Go to your course home page and the number is at the end of the url.}

\item{pattern}{Regular expression to identify registration quizzes.}
}
\value{
A data.frame
}
\description{
Collates attendance records for each student.
}
\examples{
\dontrun{
attendance <- get_course_attendance(course_id = "57351")
attendance |> group_by(name) |>
  distinct(title) |>
  separate_wider_delim(title, delim = " ", names = c("title", "date")) |>
  summarise(n = sum(!is.na(date)))
}
}
